[feature](selectdb-cloud) Add flock for meta service pid file
* Add flock when start the process
* Fix missing return when failed to listen brpc port
* Polish start/stop script

TODO: support daemon mode.

+++ cloud/script/stop_meta_service.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 cd `dirname $0`
+
+if [ ! -f meta_service.pid ]; then
+	echo "no meta_service.pid found, process may have been stopped"
+	exit -1
+fi
+
 pid=`cat meta_service.pid`
 kill ${pid}
-rm meta_service.pid
+rm -f meta_service.pid
